package refactoring for python3.12

+++ examples/iris_recognition.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
-from __future__ import print_function
-import pylab as plt
+import matplotlib.pyplot as plt
 import numpy as np
 
 # Import datasets, classifiers and performance metrics
@@ -19,23 +18,23 @@ def iris_class(value):
         return 0
 
 
-def plot_confusion_matrix(cm, title='Confusion matrix', cmap=plt.cm.Blues):
-    plt.imshow(cm, interpolation='nearest', cmap=cmap)
+def plot_confusion_matrix(cm, title="Confusion matrix", cmap=plt.cm.Blues):
+    plt.imshow(cm, interpolation="nearest", cmap=cmap)
     plt.title(title)
     plt.colorbar()
     tick_marks = np.arange(len(iris.target_names))
     plt.xticks(tick_marks, iris.target_names, rotation=45)
     plt.yticks(tick_marks, iris.target_names)
     plt.tight_layout()
-    plt.ylabel('True label')
-    plt.xlabel('Predicted label')
+    plt.ylabel("True label")
+    plt.xlabel("Predicted label")
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
 
     iris = datasets.load_iris()
 
-    viris_class = np.vectorize(iris_class, otypes=[np.int])
+    viris_class = np.vectorize(iris_class, otypes=[int])
 
     n_samples = iris.data.shape[0]
 
@@ -45,11 +44,11 @@ if __name__ == '__main__':
     n = n_samples // 3
     for i in range(0, n):
         data[j] = iris.data[i]
-        data[j+1] = iris.data[i+n]
-        data[j+2] = iris.data[i+2*n]
+        data[j + 1] = iris.data[i + n]
+        data[j + 2] = iris.data[i + 2 * n]
         target[j] = iris.target[i]
-        target[j+1] = iris.target[i+n]
-        target[j+2] = iris.target[i+2*n]
+        target[j + 1] = iris.target[i + n]
+        target[j + 2] = iris.target[i + 2 * n]
         j += 3
 
     train_data_is_the_first_half = False
@@ -65,12 +64,14 @@ if __name__ == '__main__':
         test_x = data[:n]
         test_y = target[:n]
 
-    model = Regressor(ref_functions='linear_cov',
-                      feature_names=iris.feature_names,
-                      criterion_minimum_width=5,
-                      stop_train_epsilon_condition=0.0001,
-                      l2=0.5,
-                      n_jobs=4)
+    model = Regressor(
+        ref_functions="linear_cov",
+        feature_names=iris.feature_names,
+        criterion_minimum_width=5,
+        stop_train_epsilon_condition=0.0001,
+        l2=0.5,
+        n_jobs=4,
+    )
     model.fit(train_x, train_y)
 
     # Now predict the value of the second half:
@@ -78,7 +79,6 @@ if __name__ == '__main__':
     pred_y_row = model.predict(test_x)
 
     pred_y = viris_class(pred_y_row)
-
 
     print(model.get_selected_features_indices())
     print(model.get_unselected_features_indices())
@@ -91,20 +91,20 @@ if __name__ == '__main__':
     # Compute confusion matrix
     cm = confusion_matrix(test_y, pred_y)
     np.set_printoptions(precision=2)
-    print('Confusion matrix, without normalization')
+    print("Confusion matrix, without normalization")
     print(cm)
     ax1 = fig.add_subplot(121)
     plot_confusion_matrix(cm)
 
     # Normalize the confusion matrix by row (i.e by the number of samples
     # in each class)
-    cm_normalized = cm.astype('float') / cm.sum(axis=1)[:, np.newaxis]
-    print('Normalized confusion matrix')
+    cm_normalized = cm.astype("float") / cm.sum(axis=1)[:, np.newaxis]
+    print("Normalized confusion matrix")
     print(cm_normalized)
     ax2 = fig.add_subplot(122)
-    plot_confusion_matrix(cm_normalized, title='Normalized confusion matrix')
+    plot_confusion_matrix(cm_normalized, title="Normalized confusion matrix")
 
     model.plot_layer_error()
     plt.show()
 
-    PlotModel(model, filename='iris_model', plot_neuron_name=True, view=True).plot()
+    PlotModel(model, filename="iris_model", plot_neuron_name=True, view=True).plot()
